[kernel][riscv] add icache flush logic
In the usual places where the i and d cache need to be synchronized add
fence.i and the SBI call to sync on all of the other cores. QEMU does
not emulate this but is necessary on real hardware.

fence.i is present in basic privileged spec, so no extensions are
required.

Bug: 124333

Change-Id: I6f9fe126f765725b0c8761abed2b8c36fe671716
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914499
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>