[debugger] Use DwarfBinary for indexing.
Use the zxdb DwarfBinary abstraction for indexing instead of the raw
LLVM objects. The indexing code predated the creation of DwarfBinary
which was why it was not used.

Now that all creation of llvm::DWARFContext() goes through one place,
collapse the helper code in dwarf_context.h to that location.

There should be no behavior change.

Change-Id: Ic7165b9c9ae099d423b9ea77c17e9aac82d5ab8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929613
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Adam Barth <abarth@google.com>