[netstack3] Merge BufferIpTransportContext impls
Implement UDP's `BufferIpTransportContext` for all `I` rather than
separately for `Ipv4` and `Ipv6`. This will make it easier for a follow
up CL to squash the `BufferIpTransportContext` trait into
`IpTransportContext`.

Making this change required a fair amount of churn in the unit tests,
but overall resulted in a significant simplification to the test
fixtures. Originally, there were separate test fixtures for UDP with
and without dual-stack socket support. Now there is only one fixture
which does support dual-stack. This better aligns the test-fixtures with
the real world, since UDP is a dual-stack capable protocol.

Bug: 135142

Change-Id: I942ebc4052de0b52f88ce03df0089297b880f0d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930212
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>