[bazel-sdk] allow fuchsia_package to set its platform
This change allows a fuchsia_package to specify what platform it will be
built for. This lets us have platform specific packages. There is more
work that needs to be done here but we are introducing this change
gradually to allow adoption.

This change does not add target_compatible_with flags to the package.
This will be done later when we clean up our transitions and tests.

BUG: b/303448155
Change-Id: I3e2ff74639c1c663b0a3428c44933dd182b9c4ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930560
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>