[system-updater] Use the OTA specific PackageResolver capability
pkg-resolver's OTA specific resolution capability does not add
packages to the dynamic index, regardless of the state of the retained
index.

With this change, non-OTA package resolution is the only client of the
dynamic index, so we can replace it with open package tracking.

Change-Id: I1abc8c7886c2fd4387fa9322b4ea2fd31967377d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930032
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>