[debugger] Remove llvm::DWARFContext from indexing.
This uses the context on the DWARF binary. The index previously used a
separate context for some operations, but other operations used the one
on the DwarfBinary object. This standardizes on the DwarfBinary's one.

If we need to clear cached state after indexing, we should have a
duplicate DwarfBinary (as covered by the TODO in module_symbols_impl.cc
when the indexing code is called).

There should be no behavior change.

Change-Id: Ide89376ef21faacb24a8e5db022de7819b7a3332
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931492
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>