[devshell][ffx]: Remove auto add of remote target
Automatically adding the target if assumed in a remote environment can
have unexpected behaviour to users and can cause tests to fail if the
default target was ever set to be a remote environment.

This also breaks user expectations as if a user runs

```
ffx target remove [::1]
```

The target will be successfull removed. The user can verify this in
their ffx configuration, but the next ffx invocation will automatically
add '[::1]' as the target back, making it impossible to remove until
they unset their default device.

Bug: b/293950372
Fix: b/293950372
Change-Id: Iae8933279f6476741195f38992f9faf979c1b740
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922964
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>