[doc-checker] Enabling validating reference links
This enables checking markdown reference links which use [<label>]
to reference pages.

Strictly parsing markdown also allows reference links to use
the contents between [] to be interpreted as a reference link label.
As a result, usages of [] that do not appear to be intended as
reference links are listed as warnings. Ideally, usage of [] should
be escaped with a single backslash (\).

Bug: b/289246762
Change-Id: Id2c392a40e104b844292b73017bf402590e2ca36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930294
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>