[pkg-cache] Retained index no longer blocks dynamic index
The retained index used to prevent packages from being added to the
dynamic index, so that packages downloaded during OTA would not be
protected by the dynamic index (so that system-updater could GC them
if necessary). system-updater now has an OTA exclusive package
resolver capability that never adds packages to the dynamic index, so
this blocking behavior is no longer necessary.

This CL removes the blocking behavior so that, on builds where
ephemeral and OTA resolves are intermingled, the ephemerally resolved
packages will be protected as long as they are in use (once open
package tracking is enabled).

Removing the blocking behavior also makes the behavior of the system
easier to predict, as there is one fewer dependence on hard-to-observe
state.

Change-Id: I2d5b2d8a1c623e3e0b3a2eb8f9a25e74be5065e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932284
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>