[boot-shim] Fix GIC wrong assumption on regs
Previously there was the assumption that register banks were ordered
from low to high addresses. This is wrong as seen by the CrosVM dtb,
where the second entry in reg is actually lower in address range than
first one.

We recompute the base address and the offsets based on the decoded
registers.

Change-Id: I5bdab9a20b103a9b924f030544cf538ea26fca99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930993
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>