[ffx][golden files] Print diff when mismatched
This adds displaying the unified diff between the candidate and
golden files when there is a difference.

This improves the experience of dealing with command line golden file
changes since the change causing the difference is displayed.

Since there are also some usages that compare binary files,
a flag `binary` is added to perform the filecmp based comparison
and not display the diff.

Bug: b/303542110
Change-Id: I5b0efd3bacc2b70acedc3b54f913a43b17103e1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932258
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>