[component_manager][config] Remove duplicates on vector merges
It's unclear how Component Framework would respond to duplicate
strings, so its simpler to remove them.

The deduping process is updated to sort and then dedup instead
of using a hashmap. The main benefit of this is that the
returned vector is sorted, which gives it a deterministic
output, so it is easier to test and compare.

Change-Id: I4a3bdaa18239b3d93751a0d126e3b17361a0884e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931072
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>