[sandbox] Implement duck typing on Handle for cloning and TryInto<Open>
Implement try_clone and TryInto<Open> on Handle:

* try_clone clones channels that speak fuchsia.unknown.Cloneable
* TryInto<Open> converts channels that speak fuchsia.io.Openable

When a handle is inserted into a Dict, it's stored as a Handle type
as before - there is no attempt to determine its "original" type.

Then, when a client needs the handle to do a particular behavior,
like cloning, or converting to a different type, the Handle type
does this opportunistically.

* Add a test from fxrev.dev/886692
* Sender now implements Openable and Cloneable so Dicts that
  contain senders as Handle types can be converted to a namespace.

Bug: b/303307609
Change-Id: If26ed0f51899090ff0ca4173ad507ff4b170781a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929300
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>