[elfldltl][ld] Static TLS layout
This adds a toolkit class elfldltl::TlsLayout for doing the
static TLS layout calculations in accordance with the psABI.
This uses new a elfldltl::TlsTraits template struct specialized
for each machine to describe the relevant psABI constants.

The startup dynamic linker uses this to perform static TLS layout
for each module with a PT_TLS.  The calculated totals and the
per-module offsets are exported in the passive ABI.

So far this is only tested by examining the passive ABI.
There is not yet any support for the TLS relocations.

Bug: 128502
Change-Id: Ifa18f6fc5a71dc140c5d76b9f4f1d8a6ea64753b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929896
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alex Brachet <abrachet@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>