[routing] Create NoopVisitor
Make a generic NoopVisitor and then type alias
it to make it easier to use.

This removes a macro, and it also makes it
possible to use a generic NoopVisitor if
we don't know the type ahead of time.

Bug: b/305948828

Change-Id: I01df9ae8a86562ae5741468a8a516ddced239529
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934593
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>