[netcfg] Initial interface name policy structure
Add in a new field to be read for configuration. The field is currently
optional and is represented as a default vector of size 0 when not
provided. For future CLs implementing the use of this configuration
field, a length of 0 will sustain current netcfg naming behavior and any
other lengths will use the new configuration based naming policies.

Create a struct for the MatchingRule and NameCompositionRule that will
will be extended in future CLs.

Ensure that the initial deserialization works for an interface naming
rule that has empty fields. Test will be a base for other features to
extend.

Fixed: 135096
Change-Id: Ibcabff4872565e25439b3aad2079d2c8401ef472
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929969
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>