[aml-sdmmc] Introduce own copy of io-buffer.h/c.
ddk::IoBuffer cannot be used with DFv2. This change makes a local
copy, which is used in both the DFv1 and DFv2 modes of the driver.

(The namespace of aml_sdmmc::IoBuffer is explicitly specified for
highlighting, since ddk::IoBuffer is a well-known and widely-used
class.)

Bug: b/301003087
Test: fx test sdmmc-test
Change-Id: Ibb03fe942ab0b096064ee28393a3cd5caa71e55a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934199
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>