[clock][core] Bind to either Banjo or FIDL clockimpl protocol
Make the clock core driver use either Banjo or FIDL to communicate with
the underlying clockimpl driver. Banjo is preferred if both protocols
are available. Add a test to verify this behavior, as well as check that
FIDL calls to the clockimpl driver work.

Bug: b/299501085
Test: fx test //src/devices/clock/drivers/clock
Change-Id: I51961eed820670fa172c3177808c97f10dd5126a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933389
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>