[pkg-cache] Protect cache packages from GC unconditionally
Previously, cache packages were activated in the dynamic index on
startup and were protected from GC as long as they remained active in
the dynamic index, i.e. as long as another package with the same name
(as indicated by the meta/package file) was not resolved.

Now, cache packages are still activated in the dynamic index on
startup, but they will be protected from GC regardless of if they
remain active.

This change was foretold by, and is in accordance with, the Open
Package Tracking RFC.

Fixed: b/293332528
Change-Id: I9f8a24299056add0266d35d7675fa5360e5a0e2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/935191
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>