[superproject] Roll llvm-project/libc [libc] Enable remaining string functions on the GPU (#68346)
Summary:
We previously had to disable these string functions because they were
not compatible with the definitions coming from the GNU / host
environment. The GPU, when exporting its declarations, has a very
difficult requirement that it be compatible with the host environment as
both sides of the compilation need to agree on definitions and what's
present.

This patch more or less gives up an just copies the definitions as
expected by `glibc` if they are provided that way, otherwise we fall
back to the accepted way. This is the alternative solution to an
existing PR which instead disable's GCC's handling.

GitOrigin-RevId: e56b77f58221733b209b09e5d3ae442831a9ca10
Original-Revision: 442f90f1a022bc18442e6451a058916da7eb0fce
Roller-URL: https://ci.chromium.org/b/8766379419926474289
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib40b73d0425bcc4369598d8b76b79b41ea1920a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936633