[netstack3] Don't leak DS listeners in `connect`
Dual stack listeners bound to the IPv6 any address will exist in both
the IPv4 and IPv6 socket map. When upgrading a listening socket to a
connected socket, ensure that at the socket maps for both IP versions
are updated.

This CL introduces a refactor into the existing socket-removal path so
that it can be re-used to delete the existing socket state during
`connect`. The pattern used here is to parse and store the required
remove operation before actually applying it. This allows the `connect`
call to compute the remove when it has access to `dual_stack_context` and
apply the remove while inside of the existing `with_bound_state_mut`
closure. Note that applying the remove in a separate closure would allow
data races as it would occur under a separately acquired lock.

Bug: 135204
Change-Id: If5036df7bb8a8e5392ffe833c0f94214f7fd2300
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934132
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>