[kernel][perf] Do not clear non-asserted bits.
When servicing a performance counter interrupt, do not attempt to
clear bits which were not asserted in the first place.  Attempting to
do this in an AEMU environment seems to trip a general protection
fault, although it is not 100% why.

Change-Id: I68c81f762459dd2113e1f529da0a5773b1f62b02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937295
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: John Grossman <johngro@google.com>